Revert "Separate matrix graph construction from linear system construction (#473)"
This reverts commit 2305d4874d35a87949867ad76121aa7d35bf83f2.
