+++ include/LinearSolver.h
@@ -57,6 +57,83 @@ namespace nalu{
 class LinearSolvers;
 class Simulation;
 
+const LocalOrdinal INVALID = std::numeric_limits<LocalOrdinal>::max();
+
+using RowPointers    = typename LinSys::LocalGraph::row_map_type::non_const_type;
+using ColumnIndices  = typename LinSys::LocalGraph::entries_type::non_const_type;
+
+/** LocalGraphArrays is a helper class for building the arrays describing
+ * the local csr graph, rowPointers and colIndices. These arrays are passed
+ * to the TpetraCrsGraph::setAllIndices method. This helper class is used
+ * within nalu's TpetraLinearSystem class.
+ * See unit-tests in UnitTestLocalGraphArrays.C.
+ */
+class LocalGraphArrays {
+public:
+
+  template<typename ViewType>
+  LocalGraphArrays(const ViewType& rowLengths)
+  : rowPointers(),
+    rowPointersData(nullptr),
+    colIndices()
+  {
+    RowPointers rowPtrs("rowPtrs", rowLengths.size()+1);
+    rowPointers = rowPtrs;
+    rowPointersData = rowPointers.data();
+
+    size_t nnz = compute_row_pointers(rowPointers, rowLengths);
+    colIndices = Kokkos::View<LocalOrdinal*,LinSysMemSpace>(Kokkos::ViewAllocateWithoutInitializing("colIndices"), nnz);
+    Kokkos::deep_copy(colIndices, INVALID);
+  }
+
+  size_t get_row_length(size_t localRow) const { return rowPointersData[localRow+1]-rowPointersData[localRow]; }
+
+  void insertIndices(size_t localRow, size_t numInds, const LocalOrdinal* inds, int numDof)
+  {
+    LocalOrdinal* row = &colIndices((int)rowPointersData[localRow]);
+    size_t rowLen = get_row_length(localRow);
+    LocalOrdinal* rowEnd = std::find(row, row+rowLen, INVALID);
+    for(size_t i=0; i<numInds; ++i) {
+      LocalOrdinal* insertPoint = std::lower_bound(row, rowEnd, inds[i]);
+      if (insertPoint <= rowEnd && *insertPoint != inds[i]) {
+        insert(inds[i], numDof, insertPoint, rowEnd+numDof);
+        rowEnd += numDof;
+      }
+    }
+  }
+
+  template<typename ViewType1, typename ViewType2>
+  static size_t compute_row_pointers(ViewType1& rowPtrs,
+                                   const ViewType2& rowLengths)
+  {
+    size_t nnz = 0;
+    auto rowPtrData = rowPtrs.data();
+    auto rowLens = rowLengths.data();
+    for(unsigned i=0, iend=rowLengths.size(); i<iend; ++i) {
+      rowPtrData[i] = nnz;
+      nnz += rowLens[i];
+    }
+    rowPtrData[rowLengths.size()] = nnz;
+    return nnz;
+  }
+
+  RowPointers rowPointers;
+  typename RowPointers::traits::data_type rowPointersData;
+  ColumnIndices colIndices;
+
+private:
+
+  void insert(LocalOrdinal ind, int numDof, LocalOrdinal* insertPoint, LocalOrdinal* rowEnd)
+  {
+    for(LocalOrdinal* ptr = rowEnd-1; ptr!= insertPoint; --ptr) {
+        *ptr = *(ptr-numDof);
+    }
+    for(int i=0; i<numDof; ++i) {
+      *insertPoint++ = ind+i;
+    }
+  }
+};
+
 /** An abstract representation of a linear solver in Nalu
  *
  *  Defines the basic API supported by the linear solvers for use within Nalu.
