+++ include/TpetraSegregatedLinearSystem.h
@@ -42,7 +42,6 @@ class Realm;
 class EquationSystem;
 class LinearSolver;
 class LocalGraphArrays;
-class CrsGraph;
 
 typedef std::unordered_map<stk::mesh::EntityId, size_t>  MyLIDMapType;
 
@@ -138,14 +137,12 @@ public:
 
   int getDofStatus(stk::mesh::Entity node);
 
-  //for unit testing
   int getRowLID(stk::mesh::Entity node) { return entityToLID_[node.local_offset()]; }
-  //for unit testing
   int getColLID(stk::mesh::Entity node) { return entityToColLID_[node.local_offset()]; }
 
-  Teuchos::RCP<LinSys::Graph>  getOwnedGraph();
-  Teuchos::RCP<LinSys::Matrix> getOwnedMatrix();
-  Teuchos::RCP<LinSys::MultiVector> getOwnedRhs();
+  Teuchos::RCP<LinSys::Graph>  getOwnedGraph() { return ownedGraph_; }
+  Teuchos::RCP<LinSys::Matrix> getOwnedMatrix() { return ownedMatrix_; }
+  Teuchos::RCP<LinSys::MultiVector> getOwnedRhs() { return ownedRhs_; }
 
   class TpetraLinSysCoeffApplier : public CoeffApplier
   {
@@ -203,16 +200,31 @@ public:
   };
 
 private:
-
-  Teuchos::RCP<CrsGraph>   crsGraph_;
-
-  //calls through to CrsGraph::buildConnectedNodeGraph()
   void buildConnectedNodeGraph(stk::mesh::EntityRank rank,
                                const stk::mesh::PartVector& parts);
 
   void beginLinearSystemConstruction();
 
   void checkError( const int /* err_code */, const char * /* msg */) {}
+
+  void compute_send_lengths(const std::vector<stk::mesh::Entity>& rowEntities,
+         const std::vector<std::vector<stk::mesh::Entity> >& connections,
+                            const std::vector<int>& neighborProcs,
+                            stk::CommNeighbors& commNeighbors);
+
+  void compute_graph_row_lengths(const std::vector<stk::mesh::Entity>& rowEntities,
+         const std::vector<std::vector<stk::mesh::Entity> >& connections,
+                                 LinSys::RowLengths& sharedNotOwnedRowLengths,
+                                 LinSys::RowLengths& locallyOwnedRowLengths,
+                                 stk::CommNeighbors& commNeighbors);
+
+  void insert_graph_connections(const std::vector<stk::mesh::Entity>& rowEntities,
+         const std::vector<std::vector<stk::mesh::Entity> >& connections,
+                                LocalGraphArrays& locallyOwnedGraph,
+                                LocalGraphArrays& sharedNotOwnedGraph);
+
+  void fill_entity_to_row_LID_mapping();
+  void fill_entity_to_col_LID_mapping();
 
   void copy_tpetra_to_stk(
     const Teuchos::RCP<LinSys::MultiVector> tpetraVector,
@@ -223,9 +235,30 @@ private:
   void copy_stk_to_tpetra(stk::mesh::FieldBase * stkField,
     const Teuchos::RCP<LinSys::MultiVector> tpetraVector);
 
+  int insert_connection(stk::mesh::Entity a, stk::mesh::Entity b);
+  void addConnections(const stk::mesh::Entity* entities,const size_t&);
   void expand_unordered_map(unsigned newCapacityNeeded);
   void checkForNaN(bool useOwned);
   bool checkForZeroRow(bool useOwned, bool doThrow, bool doPrint=false);
+
+  std::vector<stk::mesh::Entity> ownedAndSharedNodes_;
+  std::vector<std::vector<stk::mesh::Entity> > connections_;
+  std::vector<GlobalOrdinal> totalGids_;
+  std::set<std::pair<int,GlobalOrdinal> > ownersAndGids_;
+  std::vector<int> sharedPids_;
+
+  // all rows, otherwise known as col map
+  Teuchos::RCP<LinSys::Map>    totalColsMap_;
+  Teuchos::RCP<LinSys::Map>    optColsMap_;
+
+  // Map of rows my proc owns (locally owned)
+  Teuchos::RCP<LinSys::Map>    ownedRowsMap_;
+
+  // Only nodes that share with other procs that I don't own
+  Teuchos::RCP<LinSys::Map>    sharedNotOwnedRowsMap_;
+
+  Teuchos::RCP<LinSys::Graph>  ownedGraph_;
+  Teuchos::RCP<LinSys::Graph>  sharedNotOwnedGraph_;
 
   Teuchos::RCP<LinSys::Matrix> ownedMatrix_;
   Teuchos::RCP<LinSys::MultiVector> ownedRhs_;
