--- reg_tests/xml/matches_ml_default_2.xml
@@ -1,25 +1,0 @@
-<ParameterList name="MueLu">
-  <Parameter        name="verbosity"                        type="string"   value="none"/>
-  <Parameter        name="coarse: max size"                 type="int"      value="1000"/>
-
-  <Parameter        name="smoother: type"                   type="string"   value="CHEBYSHEV"/>
-  <ParameterList    name="smoother: params">
-     <Parameter name="chebyshev: degree"                    type="int"      value="4"/>
-     <Parameter name="chebyshev: ratio eigenvalue"          type="double"   value="20"/>
-     <Parameter name="chebyshev: min eigenvalue"            type="double"   value="1.0"/>
-     <Parameter name="chebyshev: zero starting solution"    type="bool"     value="true"/>
-     <Parameter name="chebyshev: eigenvalue max iterations" type="int"      value="15"/>
-  </ParameterList>
-
-  <Parameter        name="aggregation: type"                type="string"   value="uncoupled"/>
-<!--
-  "aggregation: drop tol" (aka aggregation thresholding) is problem dependent.
--->
-  <Parameter        name="aggregation: drop tol"            type="double"   value="0.0000000000000001"/>
-
-  <Parameter        name="repartition: enable"              type="bool"     value="true"/>
-  <Parameter        name="repartition: min rows per proc"   type="int"      value="1000"/>
-  <Parameter        name="repartition: start level"         type="int"      value="2"/>
-  <Parameter        name="repartition: max imbalance"       type="double"   value="1.327"/>
-  <Parameter        name="repartition: partitioner"         type="string"   value="zoltan2"/>
-</ParameterList>
