+++ src/HeatCondEquationSystem.C
@@ -867,7 +867,7 @@ void
 HeatCondEquationSystem::initialize()
 {
   solverAlgDriver_->initialize_connectivity();
-  //linsys_->finalizeLinearSystem();
+  linsys_->finalizeLinearSystem();
 }
 
 //--------------------------------------------------------------------------
@@ -897,7 +897,7 @@ HeatCondEquationSystem::reinitialize_linear_system()
 
   // initialize
   solverAlgDriver_->initialize_connectivity();
-  //linsys_->finalizeLinearSystem();
+  linsys_->finalizeLinearSystem();
 }
 
 void
