+++ src/LowMachEquationSystem.C
@@ -2385,6 +2385,7 @@ void
 MomentumEquationSystem::initialize()
 {
   solverAlgDriver_->initialize_connectivity();
+  linsys_->finalizeLinearSystem();
 
   // Set flag to extract diagonal if the user activates it in input file
   extractDiagonal_ = (realm_.solutionOptions_->tscaleType_ == TSCALE_UDIAGINV);
@@ -2425,6 +2426,7 @@ MomentumEquationSystem::reinitialize_linear_system()
 
   // initialize new solver
   solverAlgDriver_->initialize_connectivity();
+  linsys_->finalizeLinearSystem();
 }
 
 
@@ -3608,6 +3610,7 @@ ContinuityEquationSystem::initialize()
   }
 
   solverAlgDriver_->initialize_connectivity();
+  linsys_->finalizeLinearSystem();
 }
 
 //--------------------------------------------------------------------------
@@ -3637,6 +3640,7 @@ ContinuityEquationSystem::reinitialize_linear_system()
 
   // initialize
   solverAlgDriver_->initialize_connectivity();
+  linsys_->finalizeLinearSystem();
 }
 
 //--------------------------------------------------------------------------
