+++ src/MassFractionEquationSystem.C
@@ -619,7 +619,7 @@ void
 MassFractionEquationSystem::initialize()
 {
   solverAlgDriver_->initialize_connectivity();
-  //linsys_->finalizeLinearSystem();
+  linsys_->finalizeLinearSystem();
 }
 
 //--------------------------------------------------------------------------
