+++ src/MixtureFractionEquationSystem.C
@@ -845,7 +845,7 @@ void
 MixtureFractionEquationSystem::initialize()
 {
   solverAlgDriver_->initialize_connectivity();
-  //linsys_->finalizeLinearSystem();
+  linsys_->finalizeLinearSystem();
 }
 
 //--------------------------------------------------------------------------
@@ -875,7 +875,7 @@ MixtureFractionEquationSystem::reinitialize_linear_system()
 
   // initialize
   solverAlgDriver_->initialize_connectivity();
-  //linsys_->finalizeLinearSystem();
+  linsys_->finalizeLinearSystem();
 }
 
 //--------------------------------------------------------------------------
