+++ src/mesh_motion/MeshDisplacementEquationSystem.C
@@ -454,6 +454,7 @@ void
 MeshDisplacementEquationSystem::initialize()
 {
   solverAlgDriver_->initialize_connectivity();
+  linsys_->finalizeLinearSystem();
 }
 
 //--------------------------------------------------------------------------
