+++ unit_tests/UnitTestLocalGraphArrays.C
@@ -1,7 +1,6 @@
 #include <gtest/gtest.h>
 
 #include <KokkosInterface.h>
-#include <CrsGraphHelpers.h>
 #include <LinearSolver.h>
 
 #include <limits>
