+++ unit_tests/UnitTestTpetraHelperObjects.h
@@ -5,8 +5,6 @@
 
 #include "AssembleElemSolverAlgorithm.h"
 #include "AssembleFaceElemSolverAlgorithm.h"
-#include "CrsGraphHelpers.h"
-#include "CrsGraph.h"
 #include "TpetraLinearSystem.h"
 #include "EquationSystem.h"
 #include "kernel/Kernel.h"
@@ -37,19 +35,10 @@ struct TpetraHelperObjectsBase {
     realm(naluObj->create_realm(realmDefaultNode, "multi_physics", false)),
     eqSystems(realm),
     eqSystem(eqSystems),
-    crsgraph(Teuchos::rcp(new sierra::nalu::CrsGraph(realm, numDof)))
+    linsys(new sierra::nalu::TpetraLinearSystem(realm, numDof, &eqSystem, nullptr))
   {
     realm.metaData_ = &bulk.mesh_meta_data();
     realm.bulkData_ = &bulk;
-    realm.bulkData_ = &bulk;
-    if (numDof == 1)
-      //realm.scalarGraph_ = Teuchos::rcpFromRef(*crsgraph);
-      realm.scalarGraph_ = crsgraph;
-    else
-      //realm.systemGraph_ = Teuchos::rcpFromRef(*crsgraph);
-      realm.systemGraph_ = crsgraph;
-
-    linsys = new sierra::nalu::TpetraLinearSystem(realm, numDof, &eqSystem, nullptr);
     eqSystem.linsys_ = linsys;
   }
 
@@ -195,9 +184,7 @@ struct TpetraHelperObjectsBase {
   sierra::nalu::Realm& realm;
   sierra::nalu::EquationSystems eqSystems;
   sierra::nalu::EquationSystem eqSystem;
-  Teuchos::RCP<sierra::nalu::CrsGraph> crsgraph;
   sierra::nalu::TpetraLinearSystem* linsys;
-  //sierra::nalu::CrsGraph* crsgraph;
 };
 
 struct TpetraHelperObjectsElem : public TpetraHelperObjectsBase {
@@ -214,10 +201,7 @@ struct TpetraHelperObjectsElem : public TpetraHelperObjectsBase {
 
   virtual void execute()
   {
-    crsgraph->buildNodeGraph(realm.interiorPartVec_);
-    crsgraph->buildElemToNodeGraph({&realm.metaData_->universal_part()});
-    crsgraph->buildFaceElemToNodeGraph(realm.bcPartVec_);
-    linsys->buildElemToNodeGraph({&realm.metaData_->universal_part()}); //just to call beginLinearSystemConstruction()
+    linsys->buildElemToNodeGraph({&realm.metaData_->universal_part()});
     linsys->finalizeLinearSystem();
     assembleElemSolverAlg->execute();
     for (auto kern: assembleElemSolverAlg->activeKernels_)
@@ -275,10 +259,7 @@ struct TpetraHelperObjectsEdge : public TpetraHelperObjectsBase {
   virtual void execute() override
   {
     ThrowRequire(edgeAlg != nullptr);
-    crsgraph->buildNodeGraph(realm.interiorPartVec_);
-    crsgraph->buildElemToNodeGraph({&realm.metaData_->universal_part()});
-    crsgraph->buildFaceElemToNodeGraph(realm.bcPartVec_);
-    linsys->buildEdgeToNodeGraph({&realm.metaData_->universal_part()}); //just to call beginLinearSystemConstruction()
+    linsys->buildEdgeToNodeGraph({&realm.metaData_->universal_part()});
     linsys->finalizeLinearSystem();
 
     edgeAlg->execute();
